home_work_3/requirements.txt
simpleeval==0.9.8
python-dateutil==2.8.0